debugging

+++ model/model.py
@@ -315,7 +315,7 @@ class GLMFModelForCausalLM(GLMFModel, GenerationMixin):
             ), f"Shape mismatch in assignment: graph embedding shape {graph_embeds.shape}, input embedding shape: {inputs_embeds.shape}, graph_token_index: {len(graph_token_index)}!"
 
             inputs_embeds[0, graph_token_index[0] : (graph_token_index[-1] + 1), :] = (
-                graph_embeds
+                graph_embeds.to(inputs_embeds.dtype)
             )
         else:
             if self.is_training:
